+++ src/colony/libs/__init__.py
@@ -93,7 +93,7 @@ from .host_util import get_hostname, get_hostname_local, get_address_ip4, get_ad
 from .import_util import reload_import
 from .lazy_util import LazyClass, LazyIteratorClass, is_lazy, Lazy, LazyIterator
 from .list_util import list_intersect, list_extend, list_no_duplicates
-from .logging_util import getLogger, getLevelName, DummyLogger, StreamHandler, Formatter
+from .logging_util import getLogger, getLevelName, getLevelInt, DummyLogger, StreamHandler, Formatter
 from .map_util import map_clean, map_get, map_copy, map_copy_deep, map_duplicate,\
     map_remove, map_extend, map_flatten, map_check_parameters, map_get_value_cast,\
     map_get_values, map_output, map_normalize