+++ src/colony_wsgi.py
@@ -121,7 +121,7 @@ rewrite_path = rewrite_path and os.path.normpath(rewrite_path)
 # configuration support, defaulting to the default level in case the
 # value is not provided through configuration
 level = colony.conf("LEVEL", logging.INFO)
-level = colony.getLevelName(level)
+level = colony.getLevelInt(level)
 
 # retrieves the complete set of configuration variables associated
 # with the various paths to be used by colony, these are going to
@@ -129,6 +129,12 @@ level = colony.getLevelName(level)
 meta_paths = colony.conf("META_PATH", [], cast = list)
 plugin_paths = colony.conf("PLUGINS_PATH", [], cast = list)
 plugin_paths = colony.conf("PLUGIN_PATH", plugin_paths, cast = list)
+
+# checks if the default meta and plugins included in the colony folder
+# should be loaded or not, not loading this paths is considered non
+# standard and should only be used for special situations
+default_meta = colony.conf("DEFAULT_META", True, cast = bool)
+default_plugins = colony.conf("DEFAULT_PLUGINS", True, cast = bool)
 
 # tries to retrieve the configuration file from the environment
 # variable associated in case it fails uses the default configuration
@@ -166,7 +172,7 @@ _plugin_paths = []
 # iterates over each of the plugin paths to resolve them using
 # the glob based approach then "takes" the final list into a
 # final step of absolute path normalization
-for plugin_path in colony_configuration.plugin_path_list:
+for plugin_path in colony_configuration.plugin_path_list if default_plugins else []:
     plugin_paths.append(os.path.join(manager_path, plugin_path))
 for plugin_path in plugin_paths: _plugin_paths += glob.glob(plugin_path)
 plugin_paths = [os.path.abspath(plugin_path) for plugin_path in _plugin_paths]
@@ -174,7 +180,7 @@ plugin_paths = [os.path.abspath(plugin_path) for plugin_path in _plugin_paths]
 # iterates over each of the meta paths to resolve them using
 # the glob based approach then "takes" the final list into a
 # final step of absolute path normalization
-for meta_path in colony_configuration.meta_path_list:
+for meta_path in colony_configuration.meta_path_list if default_meta else []:
     meta_paths.append(os.path.join(manager_path, meta_path))
 for meta_path in meta_paths: _meta_paths += glob.glob(meta_path)
 meta_paths = [os.path.abspath(meta_path) for meta_path in _meta_paths]