+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-*
+* Support for new `config/python/singleton.py` python configuration mode
+* New `getLevelInt()` method to the `logging` module
 
 ### Changed